Backend/requirement.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-genai
requests>=2.31.0
pydantic==2.12.5
matplotlib
sqlalchemy
osmnx
pytrends
serpapi
httpx
urllib3==1.26.16
bcrypt==4.0.1
passlib[bcrypt]